// packages/backend/src/file-system/file-tree/file-tree.service.d.ts
export declare class FileTreeService {
    private readonly prisma;
    private readonly storageManager;
    private readonly storageService;
    private readonly storageInfoService;
    private readonly logger;
    constructor(prisma: DatabaseService, storageManager: StorageManager, storageService: StorageService, storageInfoService: StorageInfoService);
    createFileNode(options: {
        name: string;
        fileHash: string;
        size: number;
        mimeType: string;
        extension: string;
        parentId: string;
        ownerId: string;
        sourceFilePath?: string;
        sourceDirectoryPath?: string;
        skipFileCopy?: boolean;
    }): Promise<PrismaFileSystemNode>;
// ...
}